+++ qf_apps/qf_vr_prototype_app/IAR_Project/qf_vr_prototype_app.ewt
@@ -1186,6 +1186,108 @@
         </settings>
     </configuration>
     <group>
+        <name>App</name>
+        <group>
+            <name>qf_vr_prototype_app</name>
+            <group>
+                <name>fsm</name>
+                <file>
+                    <name>$PROJ_DIR$\..\fsm\fsm.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\fsm\fsm_cli_table.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\fsm\fsm_tables.h</name>
+                </file>
+            </group>
+            <group>
+                <name>inc</name>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\Bootconfig.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\FreeRTOSConfig.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\FreeRTOSFATConfig.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\Fw_global_config.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\Fw_task.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\host_interface.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\RtosTask.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\s3_host_proto_defs.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\s3x_pwrcfg.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\sec_debug.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\vr_engine_api.h</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\inc\ww_metadata.h</name>
+                </file>
+            </group>
+            <group>
+                <name>src</name>
+                <file>
+                    <name>$PROJ_DIR$\..\src\arm_rfft_fast_init_f32.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\c_lib_retarget.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\exceptions.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\hif_command_table.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\host_interface.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\main.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\main_dbg_cli_menu.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\qf_hardwaresetup.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\ql_vr_app.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\ql_vr_pe_generic.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\s3x_pwrcfg.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\sec_debug.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\vr_engine_api.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\src\watchdog.c</name>
+                </file>
+            </group>
+        </group>
+    </group>
+    <group>
         <name>BSP</name>
         <group>
             <name>quickfeather</name>
@@ -2386,108 +2488,6 @@
             <file>
                 <name>$PROJ_DIR$\..\..\..\Libraries\Amazon\Readme.txt</name>
             </file>
-        </group>
-    </group>
-    <group>
-        <name>qf_apps</name>
-        <group>
-            <name>qf_vr_prototype_app</name>
-            <group>
-                <name>fsm</name>
-                <file>
-                    <name>$PROJ_DIR$\..\fsm\fsm.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\fsm\fsm.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\fsm\fsm_cli_table.c</name>
-                </file>
-            </group>
-            <group>
-                <name>inc</name>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\Bootconfig.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\FreeRTOSConfig.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\FreeRTOSFATConfig.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\Fw_global_config.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\Fw_task.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\host_interface.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\RtosTask.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\s3_host_proto_defs.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\s3x_pwrcfg.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\sec_debug.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\vr_engine_api.h</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\inc\ww_metadata.h</name>
-                </file>
-            </group>
-            <group>
-                <name>src</name>
-                <file>
-                    <name>$PROJ_DIR$\..\src\arm_rfft_fast_init_f32.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\c_lib_retarget.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\exceptions.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\hif_command_table.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\host_interface.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\main.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\main_dbg_cli_menu.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\qf_hardwaresetup.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\ql_vr_app.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\ql_vr_pe_generic.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\s3x_pwrcfg.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\sec_debug.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\vr_engine_api.c</name>
-                </file>
-                <file>
-                    <name>$PROJ_DIR$\..\src\watchdog.c</name>
-                </file>
-            </group>
         </group>
     </group>
     <group>
